// src/app/administration/shared/models/editSuppliers.ts
export interface EditSuppliers {
    supplier_id: number;
    companyName: string;
    contactFname: string;
    contactlname: string;
    contactTitle: string;
    address: string;
    city: string;
    postalCode: string;
    country: string;
    phone: string;
    email: string;
    customerId: number;
}